Support for reserved words as columns
Insert failed when a word such as order was used as a column name.
Added [] around the column name.

Also modified the test for InsertGetUpdate().  The first Get is now a
User instead of IUser.  There is some bug where using the interface
twice first on a null result and then on the real result causes the
reader to not be open when trying to map the type.